avancée du projet

+++ src/Controller/ProjectController.php
@@ -2,50 +2,82 @@
 
 namespace App\Controller;
 
-use App\Entity\Project as EntityProject;
+use App\Entity\Project;
 use App\Form\ProjectType;
+use App\Service\FileUploader;
 use Doctrine\ORM\EntityManager;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\Routing\Annotation\Route;
 
-class ProjectController extends AbstractController{
-    #[Route("/project/create", name:"create")]
-    function create(Request $request, ManagerRegistry $doctrine) : Response{
+class ProjectController extends AbstractController
+{
+    #[Route("/project/create", name: "create")]
+    function create(Request $request, ManagerRegistry $doctrine, FileUploader $fileUploader): Response
+    {
         $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
-        $project =  new EntityProject();
+        $project =  new Project();
         $form = $this->createform(ProjectType::class, $project);
         $form->handleRequest($request);
-        if($form->isSubmitted() && $form->isValid()){
-            $entityManager=$doctrine->getManager();
+        if ($form->isSubmitted() && $form->isValid()) {
+            /** @var UploadedFile $brochureFile */
+            $brochureFile = $form->get('brochure')->getData();
+            if ($brochureFile) {
+                $brochureFileName = $fileUploader->upload($brochureFile);
+                $project->setBrochureFilename($brochureFileName);
+            }
+            $entityManager = $doctrine->getManager();
             $entityManager->persist($project);
             $entityManager->flush();
         }
         return $this->render('project/form.html.twig', [
-            "form"=> $form->createView()   
-        ]); 
-    }
-
-    #[Route("/project/edit/{id}", name:"edit")]
-    function edit(Request $request, ManagerRegistry $doctrine, EntityProject $project) : Response {
-        $form = $this->createform(ProjectType::class, $project);
-
-        $form->handleRequest($request);
-        if($form->isSubmitted() && $form->isValid()){
-            $doctrine->getManager()->flush();
-        }
-        return $this->render("project/form.html.twig", [   
-            "form"=>$form->createView()
+            "form" => $form->createView()
         ]);
     }
 
-    #[Route("/project/delete/{id}", name:"delete")]
-    public function delete(ManagerRegistry $doctrine, EntityProject $project) : Response {
+    #[Route("/project/edit/{id}", name: "edit")]
+    function edit(Request $request, ManagerRegistry $doctrine, Project $project, FileUploader $fileUploader): Response
+    {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $form = $this->createform(ProjectType::class, $project);
+       
+
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $brochureFile = $form->get('brochure')->getData();
+            if ($brochureFile) {
+                $brochureFileName = $fileUploader->upload($brochureFile);
+                $project->setBrochureFilename($brochureFileName);
+            }
+            $doctrine->getManager()->flush();
+            return $this->redirectToRoute('work');
+        }
+        return $this->render("project/form.html.twig", [
+            "form" => $form->createView()
+        ]);
+    }
+
+    #[Route("/project/delete/{id}", name: "delete")]
+    public function delete(ManagerRegistry $doctrine, Project $project): Response
+    {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $entityManager = $doctrine->getManager();
         $entityManager->remove($project);
         $entityManager->flush();
-        return $this->redirectToRoute("homepage");
+        return $this->redirectToRoute("work");
+    }
+
+    #[Route("/project/read", name: "work")]
+    public function read(ManagerRegistry $doctrine): Response
+    {
+        $repository = $doctrine->getRepository(Project::class);
+        $projects = $repository->findAll();
+        return $this->render("project/Work.html.twig", [
+            "projects" => $projects
+        ]);
     }
 }
